+++ src/graphon/engine/container_handler/builtin/iteration.py
@@ -27,11 +27,11 @@ from graphon.runtime.container_state import (
     IterationRunState,
 )
 from graphon.runtime.execution import ROOT_FRAME_ID
+from graphon.runtime.ready_queue import ResumeTask
 from graphon.runtime.runtime_state import RuntimeState
 from graphon.variables.segments import NoneSegment, SerializableSegment
 
 from ...frame import ExecutionFrame, FrameRegistry
-from ...ready_queue import ResumeTask
 
 
 @final
