+++ src/graphon/engine/ready_queue/__init__.py
@@ -1,14 +1,11 @@
-"""Ready queue implementations and serialized state helpers for Engine."""
+"""Compatibility exports for the runtime-owned ready queue."""
 
-from graphon.runtime.ready_queue import ReadyQueue
+from graphon.runtime.ready_queue import (
+    InMemoryReadyQueue,
+    ReadyQueue,
+    ReadyTask,
+    ResumeTask,
+    StartTask,
+)
 
-from .entities import ReadyTask, ResumeTask, StartTask
-from .in_memory import InMemoryReadyQueue
-
-__all__ = [
-    "InMemoryReadyQueue",
-    "ReadyQueue",
-    "ReadyTask",
-    "ResumeTask",
-    "StartTask",
-]
+__all__ = ["InMemoryReadyQueue", "ReadyQueue", "ReadyTask", "ResumeTask", "StartTask"]
