+++ src/graphon/nodes/llm/__init__.py
@@ -1,3 +1,7 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from .entities import (
     LLMNodeChatModelMessage,
     LLMNodeCompletionModelPromptTemplate,
@@ -5,7 +9,9 @@ from .entities import (
     ModelConfig,
     VisionConfig,
 )
-from .node import LLMNode
+
+if TYPE_CHECKING:
+    from .node import LLMNode
 
 __all__ = [
     "LLMNode",
@@ -15,3 +21,12 @@ __all__ = [
     "ModelConfig",
     "VisionConfig",
 ]
+
+
+def __getattr__(name: str) -> type[LLMNode]:
+    if name == "LLMNode":
+        from .node import LLMNode  # ruff: ignore[import-outside-top-level]
+
+        return LLMNode
+    msg = f"module {__name__!r} has no attribute {name!r}"
+    raise AttributeError(msg)
